Add tax calculation handler
